move nanosleep() into a header file
and re-generate .i files.
