+++ c/busybox-1.22.0/sleep_true-no-overflow_true-valid-memsafety.c
@@ -247,28 +247,6 @@ static signed long int safe_write(signed int fd, const void *buf, unsigned long 
   return n;
 }
 
-int nanosleep (const struct timespec *__requested_time,
-               struct timespec *__remaining)
-{
-        if (__VERIFIER_nondet_int()) {
-                __remaining->tv_sec = 0;
-                __remaining->tv_nsec = 0;
-                return 0;
-        } else {
-                long tv_nsec = __VERIFIER_nondet_long();
-                time_t tv_sec = __VERIFIER_nondet_long();
-                __VERIFIER_assume(tv_nsec >= 0 &&
-                                  tv_nsec <= __requested_time->tv_nsec);
-                __VERIFIER_assume(tv_sec >= 0 &&
-                                  tv_sec <= __requested_time->tv_sec);
-                __remaining->tv_sec = __requested_time->tv_sec - tv_sec;
-                __remaining->tv_nsec = __requested_time->tv_nsec - tv_nsec;
-                *bb_errno = __VERIFIER_nondet_int();
-                __VERIFIER_assume(*bb_errno != 0);
-                return -1;
-        }
-}
-
 // file coreutils/sleep.c line 52
 void syslog(int priority, const char *format, ...)
 {
@@ -487,4 +465,5 @@ inval:
   bb_error_msg_and_die("invalid number '%s'", numstr);
 }
 
+#include "busybox_sv_comp-nanosleep.h"
 #include "busybox_sv_comp_impl.h"
